Update in order to make python3 available.
This is a start for #2 issue, not a fully garanty that Python 3 works with the wrapper.
It doesn't break python2 use !
Mainly replaced the print calls to be compliant with Python 3.
Update the cmake scripts to make them search python3/cython3 using more standard cmake commands and fixed some issues by the way.
[skip ci]