Fix bug in Faust::Transform<FPP,Cpu>::updateNonZeros().
If the function were called several times in a row the value would have been as many times too bigger. Use rather the temporary variable (previously not used, just declared) to calculate the non-zeros count and then override the totalNonZeros attribute. This bug is minor because normally we don't have to call it several times if a Transform is immutable.