Simplify the C++ impl. of Faust slicing and fix a mem. leak.
The mem. leak were left in the code in waiting of an optimization.
The simplification is made of a less lazy instantiation of sliced Transform object but a much more easy to maintain/read code. For a reliability purpose the function Faust::TransformHelper::eval_sliced_transform() is now private (it must be only used by the slicing constructor).
Reminder about optimization: the slicing op. just copies parts of the first or/and last factors/matrices of the Faust. The intermediary factors are the same (not copied in memory) for the new Faust::Transform created.