Add gitlab ci jobs for windows pip-wheel pkg generation and a custom setup.py.
The setup.py is more complicated now because we need a separate windows code.
Indeed, windows wrapper dynamic libs are pre-compiled externally without setup-tools so we simply include them as data.
(I haven't been able to compile the py. wrapper through visual studio so far and setup.py forces to use it, even with --compiler option it told me it's not able to compile with g++ compiler).

Related issue: #73