Fix Faust.normalize() for the case: normalizing rows with 1-norm or inf-norm.
The C++ code is only able to normalize on columns. That's why the wrappers (in their Faust.normalize()) do a transpose when user asks for a normalization based on rows.
The bug was occurring because the transpose must be accompanied with a change of used norm in case of 1-norm and inf-norm based normalizations ;  1-norm asked -> inf-norm used, inf-norm asked -> 1-norm used behind in c++ core.
Updating unit tests in consequence (test_FaustPy.py).

Not related minor change: change GIT_STRATEGY for ctest_python job (doing a checkout now)