Support op+F, op-F, op*F, op@F and norm(F,ord), dot(F,op) in pyfaust with op a scalar or a numpy.ndarray.
Before, only F+op, F-op, etc. were supported (only in this order).
It does a lot to solve issue #104.
Other numpy operations to be supported outside of Faust class (non-member functions trying to perform the operation or relay it to numpy like pyfaust.norm/dot() do).
Updated the related jupyter notebook to precise the permutation between F and op is possible in py3.