Set the experimental mode for revision packages and exclude all experimental code (and documentation) from the release packages.
The pyfaust.fact module and mafaust.fact namespace are completely removed except for eigtj.

The doxygen doc generation is made from a copy of original source (tmp_gendoc_matfaust made by wrapper/matlab/CMakeLists.txt) to workaround a bug when it's done to on the package filtered source code (the copy is needed because we need to filter out experimental code but not modifying the original source code) (only matlab wrapper is concerned).
Manage the deletion of factparams folder in CMake script also for matfaust only.