Fix a bug in Parallel Truncated Jacobi algorithm (GivensFGFTParallel/Complex).
The bug occured in certain conditions (when the number of Givens rotations per factor -- t -- is at its max or near). An error in the break condition of a while loop provoked an exit before that the factor is full of Givens as expected.
Minor other modifs: change in the debug output/tests.