Add the abstract class Faust::GivensFGFTGen parent of all classes GivensFGFT* (complex and float sub-classes). Main purpose: refactor the code and morever avoid duplicates.
WIP: only GivensFGFTComplex is refactored in this commit.
The pyfaust and matfaust wrappers are adjusted to use a real eigenvalue vector (instead of the useless complex type when the input matrix is complex hermitian -- useless because the algorithms are only able to compute real eigenvalues). This fact is used to refactor the code (because again real and complex algos compute real eigenvalues -- same functionality/types).