Fix in PALM4MSA 2020 1) the float MEG factorization warning/error 2) NaN in the numerator and denominator of lambda.
the 1) error was bypassed through NO_LAMBDA_ERROR env. var. but is no longer necessary by comparing the S norm to zero and not to espilon anymore.
For 2) if it happens the error is raised (except if NO_LAMBDA_ERROR is used, but the algorithm can't recover this, so that's not advisable).