Fix _palm4msa_fgft (unit test green).
The bug came from the init_init_D function which in certain conditions didn't return any np.array as it should. The function was renamed set_init_D and refactored as a static member function of ParamsPalmMSAFGFT (which is more appopriate because it is used only in the class ctor).