Stop passing transpose flag argument from matfaust.Faust.mtimes* to C++ core.
mtimes_trans wasn't actually used (in fact F.'*M means calling transpose
and then mtimes) but it is left for not breaking some tests.
Faust.mtimes_trans was emptied and now depends on mtimes (it was the reverse before) which is called after a simple transpose F.' if the transpose bool is true.