Simplify the CUDA install note for win10.
What was working for CUDA 9 is not working anymore for CUDA 11 and 12.
I mean that there was a minimal install for CUDA 9 but now it
is more complicated to determine which components are really
necessary to make pyfaust work with CUDA 11 and 12 (hence the simplest
way to proceed is following the express install).