Use default compiler (gcc) for running tests (ctest).
The reason is a strange OpenMP issue in Eigen occurring with the python tests which provoke a kind of dead lock in palm4msa 2020 (it doesn't happen when compiling with gcc instead of clang, nor with the release package (compiled with clang though). It's probably due to VM specific config).