Fix a segfault in EigTJParallelSparse C++ test.
The bug was due to a buffer over-read in `choices' MATIO variable because of a wrong hard-coded value of J (used next as size of choices).
Now the dimensions are dynamically obtained from the .mat file.
Other change: use the proper function Mat_VarFree instead of system free() for freeing the variable.